abstract components out and cleanup
