+++ src/components/toolbox/inputs/index.js
@@ -1,9 +1,11 @@
 import Input from './input';
 import Textarea from './textarea';
 import AutoResizeTextarea from './AutoResizeTextarea';
+import InputWithDropdown from './inputWithDropdown';
 
 export {
   Input,
   Textarea,
   AutoResizeTextarea,
+  InputWithDropdown,
 };
