+++ src/components/toolbox/inputs/input.css
@@ -179,3 +179,27 @@
   letter-spacing: 5px !important;
   font-size: 20px !important;
 }
+
+.inputWithDropdown {
+  border: 1px solid var(--color-platinum);
+  border-radius: 3px;
+  display: flex;
+  flex: 1;
+
+  &:focus {
+    border: 1px solid blue;
+  }
+
+  &:last-child {
+    margin-bottom: 0;
+  }
+  
+  & .inputDropdownButton {
+    height: 100%;
+    border: 0;
+  }
+}
+
+.inputDropdownInput {
+  border: 0;
+}
